update benchmark to run all the current exporters
